fix: 4165 - Refactor message builder to avoid sending empty messages
